requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.0
pymysql==1.1.0
bcrypt
pytest
requests
shortuuid
python-dotenv